refactor: 파일 이름 및 폴더 구조 컨벤션 적용
