+++ app/api/api_router.py
@@ -2,8 +2,7 @@
 
 from fastapi import APIRouter
 
-from app.api import driver_api, test_api, user_db_api
-from app.api.api_key import store_api_key_api
+from app.api import api_key_api, driver_api, test_api, user_db_api
 
 api_router = APIRouter()
 
@@ -13,4 +12,4 @@ api_router.include_router(test_api.router, prefix="/test", tags=["Test"])
 # 라우터
 api_router.include_router(driver_api.router, prefix="/driver", tags=["Driver"])
 api_router.include_router(user_db_api.router, prefix="/user/db", tags=["UserDb"])
-api_router.include_router(store_api_key_api.router, prefix="/credentials", tags=["Credentials"])
+api_router.include_router(api_key_api.router, prefix="/credentials", tags=["Credentials"])
